chore: lint deploy-scripts with the contracts' solhint rules
`.solhintignore` skipped all of `l1-contracts/deploy-scripts` except a single
file, so the scripts were never linted. This removes the exclusion and adds a
scoped `deploy-scripts/.solhint.json` — solhint merges configs from the file's
directory up to the root, so the nested one only names its overrides and cannot
drift from the root rule set.

Eight rules are switched off for scripts, all for the same reason: they are
about on-chain cost or on-chain surface, and deploy scripts are never deployed.
`no-console` (scripts are CLI tools), `gas-custom-errors` / `reason-string`
(readable revert strings beat calldata-cheap ones for an operator), the four
`gas-*` micro-optimisation rules, and `no-unused-import` — the last because
those imports are load-bearing: the scripts import contracts precisely so forge
compiles their artifacts, which is what the existing
"required to disable lints to force the compiler to compile the contracts"
comments are about.

Everything else now applies. Fixed here: duplicated imports, a global import,
an interface missing its `I` prefix, `uint` -> `uint256`, and explicit
`internal` visibility on 22 state variables and constants.

Documented per-site exceptions rather than blanket rule removals: forge-std's
lowercase `vm` cheatcode handle, the deliberately shouty `_TestOnly` /
`TESTONLY_` markers, the `CoreUpgrade_v33` / `CTMUpgrade_v33` contract names
(hardcoded as forge script targets in protocol-ops), `Create2AndTransfer`'s
`deployedAddress` (force-included in AllContractsHashes, so its getter name is a
tracked ABI), and the single-quoted JSON-RPC payloads.

Note: solhint's `--fix` is not safe on this tree. `no-unused-import` deletes the
load-bearing imports and leaves dangling commas, and `quotes` rewrites the inner
`"` of a JSON payload to `'` rather than escaping them, silently producing
invalid JSON. Every fix here was applied by hand.

+++ l1-contracts/deploy-scripts/InteropLibrary.sol
@@ -136,7 +136,7 @@ library InteropLibrary {
         if (unbundlerAddress != address(0)) ++length;
         if (salt != bytes32(0)) ++length;
         bytes[] memory attributes = new bytes[](length);
-        uint attributesPointer = 0;
+        uint256 attributesPointer = 0;
         if (executionAddress != address(0)) {
             attributes[attributesPointer++] = abi.encodeCall(
                 IERC7786Attributes.executionAddress,
@@ -198,7 +198,7 @@ library InteropLibrary {
         uint256 length;
         if (indirectCall) ++length;
         bytes[] memory attributes = new bytes[](length);
-        uint attributesPointer = 0;
+        uint256 attributesPointer = 0;
         if (indirectCall) {
             attributes[attributesPointer++] = abi.encodeCall(IERC7786Attributes.indirectCall, (0));
         }
